feat(web/applications): replace mock post-documents with real api (boas-553)
