+++ apps/document-storage/.env.development
@@ -13,15 +13,15 @@ PORT=3001
 # Default : 3443
 HTTPS_PORT=3443
 
-# - - - - - - - - - Swagger - - - - - - - 
+# - - - - - - - - - Swagger - - - - - - -
 
 SWAGGER_JSON_DIR=./src/server/swagger-output.json
 
 # - - - - - - - Blob Storage - - - - - -
 
 AZURE_BLOB_STORE_CONNECTION_STRING=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;BlobEndpoint=http://0.0.0.0:10000/devstoreaccount1;QueueEndpoint=http://0.0.0.0:10001/devstoreaccount1;
-AZURE_BLOB_STORE_HOST=https://localhost:10000
+AZURE_BLOB_STORE_HOST=http://127.0.0.1:10000/devstoreaccount1
 AZURE_BLOB_STORE_CONTAINER=document-service-uploads
 
-# - - - - - - - - - Swagger - - - - - - - 
+# - - - - - - - - - Swagger - - - - - - -
 SWAGGER_JSON_DIR=./src/server/swagger-output.json
